MS-32 Fix broken build
